test(e2e): isolate portable gateway observer
