+++ test/e2e/live/portable-profile-rootless-linux.test.ts
@@ -14,9 +14,8 @@ import * as importedPortableHostPreparation from "../../../src/lib/onboard/exper
 import * as importedSandboxPrebuild from "../../../src/lib/onboard/sandbox-prebuild.ts";
 import * as importedBuildContext from "../../../src/lib/sandbox/build-context.ts";
 import { test } from "../fixtures/e2e-test.ts";
-import { spawnObservedChild } from "../fixtures/observed-child-process.ts";
 import type { TestProgress } from "../fixtures/progress.ts";
-import { stripAnsi } from "./json-envelope.ts";
+import { verifyPinnedPodmanGatewayStarts } from "./portable-profile-gateway-proof.ts";
 
 const gatewayEnvModule = (
   "default" in importedGatewayEnv && importedGatewayEnv.default
@@ -95,51 +94,6 @@ async function waitForRegistry(attempt = 0): Promise<void> {
     : new Promise<void>((resolve) => setTimeout(resolve, 250)).then(() =>
         waitForRegistry(attempt + 1),
       );
-}
-
-async function verifyPinnedPodmanGatewayStarts(
-  gatewayBin: string,
-  gatewayEnv: Record<string, string>,
-  progress: TestProgress,
-): Promise<void> {
-  const child = spawnObservedChild(gatewayBin, [], {
-    activityLabel: "command: pinned OpenShell Podman gateway",
-    progress,
-    spawn: {
-      env: { ...process.env, ...gatewayEnv },
-      stdio: ["ignore", "pipe", "pipe"],
-    },
-  });
-  let output = "";
-  child.stdout?.on("data", (chunk) => {
-    output += String(chunk);
-  });
-  child.stderr?.on("data", (chunk) => {
-    output += String(chunk);
-  });
-  try {
-    const deadline = Date.now() + 15_000;
-    let driverSeenAt = 0;
-    while (Date.now() < deadline) {
-      const plainOutput = stripAnsi(output);
-      if (/configuration error|invalid \[openshell\.drivers\.podman\] table/i.test(plainOutput)) {
-        assert.fail(`Pinned OpenShell rejected the generated Podman config:\n${output}`);
-      }
-      if (child.exitCode !== null) {
-        assert.fail(
-          `Pinned OpenShell Podman gateway exited with ${String(child.exitCode)}:\n${output}`,
-        );
-      }
-      if (/Using compute driver\s+driver=podman/.test(plainOutput)) {
-        if (driverSeenAt === 0) driverSeenAt = Date.now();
-        if (Date.now() - driverSeenAt >= 2_000) return;
-      }
-      await new Promise((resolve) => setTimeout(resolve, 100));
-    }
-    assert.fail(`Pinned OpenShell did not keep the Podman driver healthy:\n${output}`);
-  } finally {
-    child.kill("SIGTERM");
-  }
 }
 
 function writeSystemctlShim(binDir: string): void {
